Avoid root exporting `DayRangeValidationProps`
