+++ packages/x-date-pickers-pro/src/internals/utils/validation/validateDateTimeRange.ts
@@ -7,7 +7,8 @@ import {
   DefaultizedProps,
 } from '@mui/x-date-pickers/internals';
 import { isRangeValid } from '../date-utils';
-import { DateTimeRangeValidationError, DateRange, DayRangeValidationProps } from '../../../models';
+import { DayRangeValidationProps } from '../../../models/dateRange';
+import { DateTimeRangeValidationError, DateRange } from '../../../models';
 
 export interface DateTimeRangeComponentValidationProps<TDate extends PickerValidDate>
   extends DayRangeValidationProps<TDate>,
